feat: implemented validation for protein change inputs
