+++ packages/core/src/domains/codesystem/hgvs/constants.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2024.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+export const HGVS_CODE_REGEX = /(Ter|Asn|Thr|Tyr|Sec|Phe|Ala|Met|Ile|Asx|Pro|Cys|His|Trp|Asp|Glu|Xaa|Gly|Val|Gln|Leu|Lys|Arg|Glx|Ser)/gi;
