+++ packages/core/src/domains/codesystem/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
+export * from './hgvs';
 export * from './types';
 export * from './utils';
