+++ packages/mtb/package.json
@@ -36,7 +36,7 @@
         "@nuxt/kit": "^3.11.1"
     },
     "devDependencies": {
-        "@nuxt/devtools": "latest",
+        "@ilingo/vuelidate": "^5.0.2",
         "@nuxt/module-builder": "^0.6.0",
         "@types/node": "^20.12.12",
         "@vuecs/form-controls": "^2.2.0",
@@ -51,8 +51,5 @@
         "vitest": "^1.6.0",
         "vue": "^3.4.27",
         "vue-chartjs": "^5.3.1"
-    },
-    "peerDependencies": {
-        "vue": "^3.4.27"
     }
 }
