+++ packages/portal/package.json
@@ -26,7 +26,6 @@
         "@dnpm-dip/rd": "^1.9.0",
         "@fortawesome/fontawesome-free": "^6.5.2",
         "@ilingo/vuelidate": "^5.0.2",
-        "@nuxt/devtools": "latest",
         "@vuecs/countdown": "^1.0.0",
         "@vuecs/form-controls": "^2.2.0",
         "@vuecs/link": "^1.0.0",
